+++ src/app/dashboard/query.base.component.ts
@@ -2,6 +2,7 @@ import {Component, Input} from '@angular/core';
 import {Variable} from '../_models/variable';
 import {DataSource} from '../_models/data.source';
 import {DashboardView} from '../_models/dashboard_view';
+import {AuthenticationService} from '../_services/authentication.service';
 
 @Component({
     selector: 'app-query-base',
@@ -14,8 +15,14 @@ export class QueryBaseComponent {
     @Input() dateRange: string;
     @Input() dateViews: DashboardView[];
     autoGroupBy = {'Hour': '10m', 'Day': '1h', 'Week': '1d', 'Month': '1d'};
+    canEdit = false;
 
-    constructor() {
+    constructor(protected authService: AuthenticationService) {
+    }
+
+    checkCanEdit() {
+        // only admin can add and edit
+        this.canEdit = this.authService.hasRole('Administrator');
     }
 
     groupByForView(chartGroupBy) {