+++ CHANGELOG.md
@@ -21,3 +21,4 @@ All notable changes to this project will be documented in this file. See [standa
 * [#36](https://github.com/upfrontsystems/izinto-angular/issues/36) handle multiple dataset labels
 * [#35](https://github.com/upfrontsystems/izinto-angular/issues/35) add decimals for rounding chart labels
 * [#38](https://github.com/upfrontsystems/izinto-angular/issues/38) add group by dropdown
+* [#42](https://github.com/upfrontsystems/izinto-angular/issues/42) fix chart date alignment